requirements.txt
bleach==6.3.0
blinker==1.9.0
click==8.3.2
colorama==0.4.6
Flask==3.1.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
python-dotenv==1.2.2
SQLAlchemy==2.0.49
typing_extensions==4.15.0
webencodings==0.5.1
Werkzeug==3.1.8
WTForms==3.2.1
